requirements.txt
diffusers>=0.30.0
transformers
accelerate
torch
torchvision
torchaudio
gradio
gradio_client
huggingface_hub
requests
edge-tts
moviepy
opencv-python
numpy
Pillow
sentencepiece
ffmpeg-python
python-dotenv
